* Fixed a compiling error.
git-svn-id: file:///usr/local/src/otland/tfssvnrepo/branches/0.2@5373 c161f52b-5830-4e7a-bee0-5dea2c17262e

+++ networkmessage.cpp
@@ -130,7 +130,7 @@ void NetworkMessage::AddItem(const Item* item)
 	AddU16(it.clientId);
 
 	if(it.stackable)
-		AddByte(std::min(255, item->getSubType()));
+		AddByte(std::min((uint16_t)255, item->getSubType()));
 	else if(it.isSplash() || it.isFluidContainer())
 	{
 		uint32_t fluidIndex = item->getSubType() % 8;
